+++ packages/bot-detector/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinyhttp/bot-detector
 
+## 0.5.35
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.5.34
+  - @tinyhttp/router@0.5.9
+
 ## 0.5.34
 
 ### Patch Changes